SCRUM-73-Integrate-Frontend-with-API-Data-Proof-of-Concept
loading.tsx: Added a loading state for the catalog. Will get called if something is still loading. Feel free to edit the spinner through loading.tsx.

page.tsx: Added state messages for error and empty. Note: empty message and parameter must be changed when handling all dynamic item cards. Fetched data for ID = 1 from the API as a proof of concept, and it can be scaled to include pagination.